requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib
psycopg2
bcrypt==4.3.0
python-jose